Move fuse-mounting code to a dedicated function
